Move key-recording helpers and fixtures to conftest module
